SetAffinityForMultiThread before execve
ckati limit the number of CPU cores to be used with
SetAffinityForSingleThread at the beginning of the program.
Since it affects program invoked by ckati, child processes are
executed with unexpectedly small number of CPUs.
Let me SetAffinityForMultiThread before execve to reset that.

Since Goma compiler_proxy is invoked by ckati on Android, it limit
the number of CPUs used for compiler_proxy's local compiler run.

Change-Id: Ib95d6610d3798cfa966063aba94164bb5df077b4
Signed-off-by: mydongistiny <jaysonedson@gmail.com>